Refactor batch processing discovery package
