+++ src/pix_framework/discovery/batch_processing/batch_characteristics.py
