+++ tests/pix_framework/discovery/batch_processing/assets/event_log_6.csv
