+++ tests/pix_framework/discovery/batch_processing/test_features_table.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
 import pandas as pd
-from pix_framework.discovery.batch_processing_discovery.features_table import _compute_features_table, _get_features
+from pix_framework.discovery.batch_processing.features_table import _compute_features_table, _get_features
 from pix_framework.io.event_log import DEFAULT_CSV_IDS
 
 assets_dir = Path(__file__).parent / "assets"
